+++ src/main/java/org/webbitserver/WebSocket.java
@@ -1,0 +1,11 @@
+package org.webbitserver;
+
+import java.util.concurrent.Future;
+
+public interface WebSocket {
+    Future<WebSocket> start();
+
+    WebSocket close();
+
+    void reconnectEvery(long millis);
+}