+++ HISTORY.md
@@ -1,3 +1,8 @@
+In Git
+==================
+
+* WebSocketClient can reconnect. (Aslak Hellesøy)
+
 0.3.6 (2012-01-30)
 ==================